Refactor settings
